CRUD Estado assistências técnicas e algumas correções
CRUD Estado das asssistências técnicas - concluído
Botão eliminar dos listar e visualizar com janelas pop-up iguais.
Técnico inserir - com validação de user/login e confirmação de password.

+++ departamento_Visualizar.php
@@ -37,7 +37,7 @@ include_once("conexao.php");
 		<div class="pull-right">
 			<a href='departamento_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
 			<a href='departamento_Editar_Formulario.php?id=<?php echo $resultado['idDepartamento']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='departamento_Eliminar.php?id=<?php echo $resultado['idDepartamento']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
+			<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='departamento_Eliminar.php?id=<?php echo $resultado['idDepartamento']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>
 		</div>
 	</div>
