+++ office_Visualizar.php
@@ -69,8 +69,8 @@ include_once("conexao.php");
 	 <div class="row">
 		<div class="pull-right">
 			<a href='office_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
-			<a href='office_Editar_Formulario.php?id=<?php echo $resultado['idOffice']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='office_Eliminar.php?id=<?php echo $resultado['idOffice']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
+			<a href='office_Editar_Formulario.php?id=<?php echo $resultado['idOffice']; ?>'><img src="imagens/edit.ico" width="30px"></a>
+				<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='office_Eliminar.php?id=<?php echo $resultado['idOffice']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>
 		</div>
 	</div>
