+++ requerente_Visualizar.php
@@ -37,7 +37,7 @@ include_once("conexao.php");
 		<div class="pull-right">
 			<a href='requerente_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
 			<a href='requerente_Editar_Formulario.php?id=<?php echo $resultado['idRequerente']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='requerente_Eliminar.php?id=<?php echo $resultado['idRequerente']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
+			<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='requerente_Eliminar.php?id=<?php echo $resultado['idRequerente']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>
 		</div>
 	</div>
