+++ software_Visualizar.php
@@ -69,9 +69,9 @@ include_once("conexao.php");
 	
 	 <div class="row">
 		<div class="pull-right">
-			<a href='software_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
-			<a href='software_Editar_Formulario.php?id=<?php echo $resultado['idSoftware']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='software_Eliminar.php?id=<?php echo $resultado['idSoftware']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
+			<a href='software_Listar.php'><img src="imagens/list.png" width="30px"></a>
+			<a href='software_Editar_Formulario.php?id=<?php echo $resultado['idSoftware']; ?>'><img src="imagens/edit.ico" width="30px"></a>
+			<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='software_Eliminar.php?id=<?php echo $resultado['idSoftware']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>
 		</div>
 	</div>
