+++ tecnico_Visualizar.php
@@ -37,7 +37,8 @@ include_once("conexao.php");
 		<div class="pull-right">
 			<a href='tecnico_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
 			<a href='tecnico_Editar_Formulario.php?id=<?php echo $resultado['idTecnico']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='tecnico_Eliminar.php?id=<?php echo $resultado['idTecnico']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
+			<!--<a href='tecnico_Eliminar.php?id=<?php echo $resultado['idTecnico']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>-->
+			<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='tecnico_Eliminar.php?id=<?php echo $resultado['idTecnico']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>			
 		</div>
 	</div>
