+++ tipo_Assistencia_Visualizar.php
@@ -36,9 +36,9 @@ include_once("conexao.php");
 	 <div class="row">
 		<div class="pull-right">
 			<a href='tipo_Assistencia_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
-			<a href='tipo_Assistencia_Editar_Formulario.php?id=<?php echo $resultado['idTipo_Assistencia']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='tipo_Assistencia_Eliminar.php?id=<?php echo $resultado['idTipo_Assistencia']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
-		</div>
+			<a href='tipo_Assistencia_Editar_Formulario.php?id=<?php echo $resultado['idTipo_Assistencia']; ?>'><img src="imagens/edit.ico" width="30px"></a>
+			<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='tipo_Assistencia_Eliminar.php?id=<?php echo $resultado['idTipo_Assistencia']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>
+	</div>
 	</div>  
 	
 	<div class="row">
