+++ tipo_Equipamento_Visualizar.php
@@ -36,8 +36,8 @@ include_once("conexao.php");
 	 <div class="row">
 		<div class="pull-right">
 			<a href='tipo_Equipamento_Listar.php'><img src="imagens/list.png" width="30px"></a></a>
-			<a href='tipo_Equipamento_Editar_Formulario.php?id=<?php echo $resultado['idTipo_Equipamento']; ?>'><img src="imagens/edit.ico" width="30px"></a></a>
-			<a href='tipo_Equipamento_Eliminar.php?id=<?php echo $resultado['idTipo_Equipamento']; ?>'><img src='imagens/edit_delete.png' width='30px'></a>
+			<a href='tipo_Equipamento_Editar_Formulario.php?id=<?php echo $resultado['idTipo_Equipamento']; ?>'><img src="imagens/edit.ico" width="30px"></a>
+			<a href="#" onclick="javascript: if (confirm('Deseja remover este registo?'))location.href='tipo_Equipamento_Eliminar.php?id=<?php echo $resultado['idTipo_Equipamento']; ?>'"><img src='imagens/edit_delete.png' width='30px'></a>
 		</div>
 	</div>
